+++ src/llm_dit/pipelines/__init__.py
@@ -43,8 +43,6 @@ from llm_dit.pipelines.generate import (
     create_video_modality,
     cleanup_memory,
 )
-from llm_dit.pipelines.ltx2_config import LTX2OptimizationConfig
-
 # WanVideoPipeline for Wan 2.1/2.2 video generation
 try:
     from llm_dit.pipelines.wan_video import WanVideoPipeline
@@ -77,5 +75,4 @@ __all__ = [
     "create_position_indices",
     "create_video_modality",
     "cleanup_memory",
-    "LTX2OptimizationConfig",
 ]